Refactoring the IP support, and fixing #1846.
I've made the IPMess stuff a lot less messy, and refactored
a lot of the util/ip module, including naming it more sensibly.

The biggest changes are that I moved the big case statement into
a constant and then used a bit of dispatch-style logic to use it,
and I eliminated a bunch of duplicate code in the ipmess.rb file.

Added some test data for FreeBSD and fixed a bug in my map
logic pointed out by Paul Nasrat.

I've also fixed #1846, in that the interface list now s/:/_/g.

Signed-off-by: Luke Kanies <luke@madstop.com>