correctly compare values - fixes #2021
this ensures we can compare all kind of objects and
not only instances of strings.
It compares strings in a case-insensitive manner and converts
symbols to strings.

introducing this behavior required that we introduce
a convert util method, to ensure that we convert the value
correctly. Introduced this method in other places as well.

This behavior change requires that we drop one test, which have
become anyway deprecated.