Fixing #1918 - facter --puppet always works
The problem was that Facter wasn't telling Puppet to read
your puppet.conf, so if you'd set vardir or libdir
in it then you didn't get the appropriate settings and thus
not know where to find the facter plugins.

This is a bit of a ham-handed approach, but it always works.

Signed-off-by: Luke Kanies <luke@madstop.com>