Fix broken solaris zone tests on EC2
This cleans up xen and vserver detection to enable us to stub out so when we
happen to be running tests on xen we don't report as that.

More cleanup is needed in this area but this should give us a green build.

This renames the tests to be consistent with current naming convention