(#5040) fact virtual should detect hpvm
With HP-UX you can build virtual machines that are often refered to as
HP-VMs.

This patch detecs HP-VMs by introducing a new function hpvm? that will
check the output of /usr/bin/getconf MACHINE_MODEL. This should not
depend on any tools that might be not installed.

If inside a HP-VM the command will say something like
ia64 hp server Integrity Virtual Machine

while on real hardware the output could be
ia64 hp server rx660

so searching for "Virtual Machine" should work.

Currently it only works if the guest is also running HP-UX.
(I guess this is the most common usecase).