(#2270) add IPv6 support to facter core.
Now that IPocalypse has happened, IPv6 support in Facter core would be nice to
have.  So, we add the appropriate code to start handling that.

The ipaddress6 fact as supplied included some smart code to try determining
the "primary" address using DNS to resolve the AAAA record for the host FQDN.

While this was smart, it actually didn't work: facter prefers the longest
confine list, so the *stupid* mechanisms that were kernel-specific would
override the smarter and more portable mechanisms.

We strip that code out for now, which also brings this into line with the
existing ipaddress fact; improving both would be good, but it should be uniform.

Paired-With: Matt Robinson <matt@puppetlabs.com>
Paired-With: Max Martin <max@puppetlabs.com>