(#2270) Remove DWIM code from ipaddress on Darwin.
This mirrors the behaviour of the newer code, and eliminates some DWIM work we
do on a single, specific platform that should, better, be done on a global
platform.

This would be reintroduced globally, applied to all platforms, or otherwise
rebuilt from a consistent base of specified behaviour.

Paired-With: Matt Robinson <matt@puppetlabs.com>
Paired-With: Max Martin <max@puppetlabs.com>