ensure that facts specified by environment variables take precedence
There was this unpleasant inconsistency:

On an OSX host:

    $ facter operatingsystem
    Darwin

    $ facter_operatingsystem=Not_Darwin facter operatingsystem
    Not_Darwin

But on a linux/solaris host:

    $ facter operatingsystem
    CentOS

    $ facter_operatingsystem=Not_CentOS facter operatingsystem
    CentOS

As the operatingsystem fact resolution for linux-based kernels has higher
precedence than the environment variable as it has more matching confines than
the value from the environment variable.

This patch adds from_environment to the resolution mechanism, which makes the
resolution have an artificially high weight by claiming the length of its confines
array is 1 billion.