(#6728) Facter improperly detects openvzve on CloudLinux systems
Update the OpenVZ fact to be slightly more resilient in it's checking of
OpenVZ, so it doesn't clash with CloudLinux's LVE container system.

It checks if /proc/vz/ actually has anything in it, which if it doesn't,
is Cloudlinux. Uses the better OpenVZ detection methodology from #2242,
which is reading /proc/self/status for the envID which OpenVZ uses.