(#4980, #6470) Fix architecture in Darwin and Ubuntu
Architecture now relies on the hardwaremodel fact unless special cased
otherwise, such as for linux distributions that require amd64 as the
expected architecture. Ubuntu was added as a special case, OpenBSD was
collapsed into the current architecture fact and Darwin was added by
removing the kernel confine statement for the architecture fact.