(#6515 and #2945) Fix processorcount for arm, sparc & ppc for linux.
Previously we were unable to check processor type and count on other
architectures for linux. This fix corrects that.

To remove complication from the fact we have moved the logic for parsing
cpuinfo and lsdev into their own static classes in Facter::Util::Processor.
This is to help with stubbing and to segregate that action as now we have
more conditional cases.

Tests and corresponding cpuinfo fixtures have been added to test those
alternative platforms as well.