(#6614) Update ipaddress6 fact to work with Ruby 1.9
Calling #to_s on an Array such as ["foo"] in Ruby 1.9 will result in
the string '["foo"]', instead of stringifying the element in the array
which would have given the expected result of "foo".  Since the
element of the array we're dealing with is already a string, we can
just grab it out of the array by using #first.

Paired-with: Josh Cooper <josh@puppetlabs.com>