(#7507) Fix 1.9.2 test failure
Dir.glob returns an array, but a test was stubbing it to return a
string.  In Ruby 1.8.7 if you call enumerable methods (each, collect,
etc) on a string, the string is split on \n first.  This meant the poor
stubbing didn't affect 1.8.7, but 1.9.2 is more strict and won't
automatically convert when you call enumerable methods on strings.

Paired-with: Josh Cooper <josh@puppetlabs.com>