(#7670) Stop preloading all facts in the application
If a requested fact isn't found in the same location as its name, we want to
load all of the facts to find it. However, to simplify that, we were previously
just preloading all the facts every time. Because requesting a fact now
implicitly loads all facts if necessary, we can rely on that, providing results
much more quickly in the case where facts do match their filenames.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>