Maint: Deprecate facter resolution interpreter parameter
Facter::Util::Resolution used to track the interpreter for the current
system, although this interpreter was never used to execute anything.
This commit removes the need to specify the interpreter parameter, and
warns once if the parameter is specified or accessed via accessors (for
backwards compatibility).

Paired-with: Jacob Helwig <jacob@puppetlabs.com>