(#8660) Fix destdir option on Windows
Specifying the --destdir option failed on windows because the install
script attempted to concatenate two absolute paths together. On Unix,
this is fine, but on Windows, it fails because the colon (part of the
drive specifier) is not a valid filename character. This commit adds a
method to join two paths, stripping off the drive specifier, if
present.

Also added the deprecation warning if the DESTDIR environment variable
is specified (to match the puppet install.rb script).

Reviewed-by: Cameron Thomas <cameron@puppetlabs.com>