(maint) Remove global var from domain and hostname
The hostname and domain facts were tightly coupled because the hostname
fact was setting a global variable that the domain fact was using.
Removed the logic from the hostname fact that was setting the global
variable and moved it into the domain fact.

Added tests to verify the behavior of the hostname facts and the domain
facts.