(#8439) Add Facter::Util::WMI module
This commit adds a WMI module with methods for connecting to the WMI
service and executing queries. This code is based on Nick's ASDI
module in puppet, but only contains the WMI calls that we need. As
part of this change, it ensures the default WMI moniker string is
consistently specified as the following:

  winmgmts:{impersonationLevel=impersonate}!//./root/cimv2

However, the wmi_resource_uri method can be overridden to provide
different behavior if necessary, e.g. different WMI namespace.

This commit also refactors the various places that we were using WMI
(inconsistently). Another motivation for this change is to allow the
WIN32OLE.connect and ExecQuery methods to be stubbed, which wasn't
possible previously.

This commit also adds manufacturer spec tests for Windows, which stub
the connect and execquery methods so that the tests can be run on
non-Windows platforms.