(#8439) Add physicalprocessorcount and processor facts on Windows
This commit adds the 'physicalprocessorcount', 'processor{n}' and
'processorcount' facts. The 'physicalprocessorcount' fact is obtained
by counting the number of Win32_Processor instances. Note that the WMI
query does a select on just the Name property, because it is faster
than doing a 'select *'

On Windows 2008, each Win32_Processor represents a physical processor,
and the NumberOfLogicalProcessors property (which includes both multi
and/or hyperthreaded cores) represents the number of logical
processors. For example, a dual-core processor, with quad-hyper
threads per core, will report 1 physical processor and 8 logical
processors.

Note that the NumberOfCores property could be used to distinguish
between multi-core and hyperthreading processors, but the fact does
not distinguish between them.

On Windows 2003, each Win32_Processor represents a logical processor,
and the NumberOfLogicalProcessors property is not available. In this
case, the physicalprocessorcount fact will be over-reported, but the
number of logical processors will be correct.

With that said, if this hotfix is installed, then 2003 behaves like
2008:

  http://support.microsoft.com/kb/932370

The Win32_Processor.Name property contains extra spaces:

  Intel(R) Core(TM) i7 CPU       M 620  @ 2.67GHz

So we 'squeeze' the output to eliminate duplicate spaces:

  processor0 => Intel(R) Core(TM) i7 CPU M 620 @ 2.67GHz