(#8437) Refactor macaddress fact on Windows
Moved the code for performing macaddress resolution into a utility module.
This is being done in preparation for supporting interface-specific
macaddresses, e.g. macaddress_<ifname> = <mac>.

As before, the 'macaddress' fact returns the macaddress of the first
network adapter on which IP is enabled.

Previously, the fact was selecting all properties, but only retrieving
the MACAddress property. This commit changes the query to only select
the property it needs, which reduces the amount of data obtained from
WMI.