(#8437) Add interface-specific ip facts for Windows
This commit adds Windows to the list of Facter::Util::IP
'supported_platforms'

It adds support for the 'interfaces' fact, e.g.

  interfaces =>
  Loopback_Pseudo_Interface_1,Local_Area_Connection,Teredo_Tunneling_Pseudo_Interface

On Windows, the name of the interface can be edited to include
non-alphanumeric characters, which are not valid fact names. This
commit changes the alphafy method to ensure the returned value only
includes alphanumeric characters (and underscore).

And the ipaddress, netmask, network, and ipaddress6 per-interface
facts, e.g.

  ipaddress_local_area_connection => 172.16.138.218
  netmask_local_area_connection => 255.255.255.0
  network_local_area_connection => 172.16.138.0

  ipaddress6_teredo_tunneling_pseudo_interface =>
  2001:0:4137:9e76:24de:36a7:53ef:7525

Note the per-interface macaddress fact is not yet supported.

Also these facts are only supported on 2008, because the output and
syntax of netsh is different on 2003. Also 2003 has dependencies on
the Routing and Remote Access service, which may not be running.