(#2157) Support for windows batch files added to parser & relevant tests updated.
So for unix systems testing if a file is executable is enough, but on
windows all files are executable. So now the ScriptParser will conditionally
change behaviour and use match_extension "bat" for windows systems.

Tests that hit ScriptParser now create .bat files with batch file
script content instead on windows so all tests on windows now pass.