(#2157) Refactor Parser class to add support for --timing for external facts.
I've refactored the way 'results' is used internally by adding a proxy
method which wraps the results call in timing code. So now you can use
--timing in facter and it will report time spent in external scripts.

Also - I've moved the caching logic into this new method ('values') which
means I can report if the timing was with cache or not.

Both of these refactors adds the benefit of moving this work out of the
parser subclasses which means less scaffolding for new parsers. It also
means all parser types can use cache with no work. Adding ttl's to the Json
parser for example actually yields a change from 50 ms to 0.05 ms so this
is very beneficial.