(#2157) Changed logic so ttl of zero now disables cache operations.
Previously a ttl of 0 was cache forever which is surprising. But other
parts of the code weren't in line with this assumption and they were
causing weird side-effects (like if you deletd the ttl file the cache
would never expire, but only if a ttl file existed first).

I've reversed that logic now. As a side effect I've had to stub handling
of ttl in our tests that involve caching which makes the tests look
a little messy but I believe the logic is more correct this way.