(#2157) Make Facter::Util::Cache interface use class methods.
To enable a global concept of a cache its far easier to work with
class methods and attributes on a class. This will enable use to
use the cache from any other object, thereby removing the requirement
to initialize the cache.

The Cache class has now been modified to act independently and
require no initialization.

The Parser and DirecterLoader classes have been modified to use this new
mechanism and all tests have been modified as well.