(#2157) Cache using ttl even if the fact returns nil.
We weren't caching nil values if the ttl was set before so this
patch fixes that.

The problem was that a call to Cache.get was returning nil for
a number of events: empty data, tty zero, tty expired and nil data.

To distinguish nil data from those other conditions Cache.get will
now raise exceptions that must be caught for: empty data, tty zero and
tty expired conditions.

Now when Cache.get returns nil it really means the data is nil.