(#9517) Fix physicalprocessorcount on windows
A broken test led to a broken fact. The WMI.execquery was incorrectly stubbed
to return an array when the actual WMI.execquery does not return an array. This
means that length, which works on arrays, does not work with WMI.execquery.
This fixes both the fact and the test. The test is unfortunately lifted to a
higher level, but it has the benefit of being correct.

Thanks to Eric Stonfer for the fact fix.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>