(#7996) Restrict solaris cpu processor detection
x86_64 based solaris machines have a pkg_core_id field in output of
kstat cpu_info, which denotes the core id of a specific core relative to
the cpu. This could cause misreporting of processor count. The regex to
count cores was restricted to prevent this. Additional x86_64 tests were
added to verify this behavior.