(#9708) Confine facts by kernel not operating system and remove confine for hardwareisa
This will allow the following facts to continue working even when new variants
of Linux, BSD, etc are supported in the future.

* lsbmajdistrelease
* macaddress
* uniqueid

For hardwareisa the confine was removed entirely. On any system that returns
"uname -p", the value should be this fact. On systems that don't have
"uname -p", the resolution will not be used. Thus, it shouldn't be necessary to
confine the resolution by kernel.

Previously we were constantly requiring patches to be added whenever a new
Linux distribution was added. These changes should reduce this need. The fact
'operatingsystem', 'osfamily' and 'operatingsystemrelease' will still require
patches however.

New spec tests were added where they were missing.