(#9555) Spec tests: Change all cases of tabs and 4 space indentation to 2 space indentation.
Since 2 space indentation seems to be Puppets (and the ruby communities)
standard this patch converts all incorrect indentation to 2 spaces.

The fact that we were mixing the indentation was causing people to mix them
within files - sometimes using 4 space, sometimes 2 space. This single change
makes it consistent across all the code.

This commit fixes any whitespace cleanups for the spec tests missed in the last commit.