(#10395) Fix manufacturer test moching for Windows
Commit b2a66a97eb518c5d6207353c18ce6056e9715995 brought in changes to
test for prtdiag returning nothing.

Because the fact returns nothing though, on windows the tests fall back
to use Windows own mechanism for detection therefore causing the test
assertion to fail. It should be nil, but instead its return a real
manufacturer value.

To stop it from falling back to the windows mechanism, I'm simply moching
the kernel fact to be 'SunOS' instead.