(#10490) Handle case where no macaddress can be found
If we do not find a macaddress when running ifconfig -a (case is
described in #10490 and is reproduceable if you deconfigure all
interfaces despite lo0) we get the error:

    Could not retrieve macaddress: private method split' called for
    nil:NilClass

Return nil as a macaddress if no macaddress can be found. The same
behaviour is already used for macaddresses on windows.