(##11041) Add dmidecode as a requirement for rpm
We were implicitly relying on dmidecode to determine certain facts
without being certain that it was installed.  This change to the rpm
spec file will ensure we have dmidecode pulled in by rpm/yum as a
dependency and thus causing the correct functionality on facter.

Note this change only impacts EL based systems.  The debian packaging is
handled in another repository.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>