(#11196) Scan all arp entries for an ec2 mac
This patch now scans all arp entries for the magic EC2 mac address. At times
the mac entry was being returned out of order and since we only looked at the
first entry there were cases where the test would fail.

It also now removes the dependency on the arp fact which has become only
important to the EC2 fact anyway. This was to avoid hacking the arp fact
(which was obviously built for a different purpose) just to fix this issue.