(#11328) Fix uptime detection on OpenBSD
Of the four or so uptime detection methods, the sysctl one is the
closest however OpenBSD lacks the -b option. Using -n instead the first
number in the output will be the epoch seconds, on OpenBSD this is all
you get, but with Darwin (and probably the other *BSD's) you get the
epoch seconds, microseconds and the pretty-printed output.

New test cases added and the old ones with the endian-specific packed
output are removed.