(#11583) Switch request method to open-uri monkey patch 'open'
This makes the code consistent with the rest of the EC2 fact code by just using
open for http requests instead of the full open_uri method. This is to make my
testing consistent and is also slightly cosmetic.

I also fixed the indentation on those two lines as well.