(#11583) Add basic coverage to the ec2 fact
The only goal here is to add coverage based on the current behaviour so we have
something to work with when other contributors work on the code.

We've only added very simple bad-case testing (ie. timeout) to match the code
at the moment, however the good-case testing covers flat & structured meta-data
and user-data. While we test both ec2 and eucalyptus cases, we only check basic
user-data for eucalyptus for now (as the code is the same in both cases).