Fixed #11666 - Add is_ec2 and is_euca facts
Sometimes you just want to know if a host is EC2 or Eucalyptus. This
adds two facts in a similar vein to the "is_virtual" fact.

The is_ec2 fact is set to true if an EC2 ARP entry is present and the
is_euca fact to true if there is a Eucalyptus MAC address.

Otherwise both default to false.