(#11566) Add windows support for ec2 facts
This patch adds support for detecting ec2 on windows. This works by modifying
the linux methodology by using arp -a instead of arp -an and searching for the
mac address with a hyphen delimiter (as apposed to a quote).

I've added tests and a sample fixtures which adds output from arp -a from a
windows machine and linux machines, on ec2 and not on ec2.

I've also re-worked the decision making into a util class so the testing is
much easier to write and work with, so now we can test the individual mechanism
for detecting that we are in a cloud on their own. This will be much better
abstracted into their own fact(s) but for now this has the least impact to
solve the problem at hand. In the future this logic (and tests) can certainly
be re-used if such a fact was evercreated.

Thanks to Feifei Jia <fjia@yottaa.com> for contributing the original code.