(#11848) Don't hard code ruby install paths in Windows batch files
Previously, the facter.bat file hard coded the path to the ruby
installation, making it impossible to move the ruby install directory.

This commit changes the script to use the `%~dp0` batch file modifier,
which resolves to the drive letter and path of the directory of the
batch file being executed.

Windows XP and later all support the `%*` modifier, so this commit
removes the Win 9x code paths that are not supported.