(#12012) Move "with_env" utility method from test code into lib code
The utility method "with_env", which was defined locally in loader_spec.rb, is actually
more generally useful (and will be used to complete the resolution of #12012).  Move it
out of the test code and into Facter::Util::Resolution so that it is accessible to
production (non-test) code.  Refactor loader_spec test accordingly.