(#12012) Remove global override of LANG environment variable
'LANG' environment variable must be set to 'C' in order to reliably parse output of certain commands when building up core Facter facts (to prevent locale differences from affecting the format of the output).  However, it does not need to happen at a global-ish scope whenever facter is 'require'd;  move this behavior from the facter.rb into the local scope of the Facter::Util::Resolution.exec method; restore original value of environment variable after executing any individual command.

Note that, while this behavior is better than the previous behavior (global override), it is still not possible for a custom facter fact to be written to expect any other locale besides 'C'.  Eventually might prefer to find a solution that allows callers to override this behavior if desired.