(#12077) Add pciutils RPM dependency
This patch adds pciutils as a dependency to the redhat spec file facter/conf/redhat.  Facter uses both lspci and dmidecode to determine the virtual/is_virtual facts/detection of hypervisors, so it should also depend on the pciutils package.  Also reference RHBZ# 783749 for addition of the same to fedora, originally noted by Dominic Cleal.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>