(#12012) Fix bug with overriding LANG environment variable
There is a yields? method called with_env which is intended to temporarily override environment variables and then restore them to their previous state after executing the caller's block of code.  The Facter::Util::Resolution.exec() method uses this, but the yield block contained several occurrences of the "return" statement; this statement breaks out of the rest of the execution of the yields? method, and thus the environment variables were never being restored by the final stanza of the with_env method.  This changeset fixes that bug.

Also adds a spec test to reproduce/verify.