(#11660) Adds feature SSHFP (#11659) Adds ssh spec
RFC4255 SSH fingerprints are now facts in the form SSHFP_RSA,SSHFP_DSA,
SSHFP_ECDSA.

This also performs SHA2 and ECDSA that are part of the draft:
http://tools.ietf.org/id/draft-os-ietf-sshfp-ecdsa-sha2

(www.iana.org/assignments/dns-sshfp-rr-parameters).

Adds test cases for the ssh fact module. Reworks a bit of the ssh fact
implementation so that the test framework can work.