(#12790) Raise an xception if recursion is detected
Facter can already detect if fact dependencies will end up in a loop.
Recursions can lead to strange errors because facter will return nil (or
an already cached value) for the fact value.

The change now changes the behaviour upon recursion detection: Dont
print a debug message but throw an error. This way a recursion is a
lot easier to detect so we can actually fix it.