(#12831) Fix recursion on first kernel fact resolution
We encounter a recursion if we want to detect the kernel fact for the first
time:

The kernel codeblock calls

    Facter::Util::Resolution.exec("uname -s")

and Facter::Util::Resolution#exec wants to detect if we can use `which`
to get the full path of the command. But the method
Facter::Util::Resolution#have_which tries to query the kernel fact again
to check if we are on windows.

Change the check in have_which so we dont have to query the kernel fact.