(#13683) Distinguish "Funtoo" from "Gentoo" Linux
As the ticket owner Marc Richter states:

Funtoo Linux is a Fork of Gentoo Linux from the former creator of Gentoo
Linux, Daniel Robbins. The two derivates behave mostly the same, but
they are differing more and more over time. So the package manager
portage (and it's frontend emerge) for example are behaving a little
different already. Packages are named different, versions differ, and so
on. It would be great if one could differ between Funtoo and Gentoo in
his manifests to install different package versions on Funtoo/Gentoo or
package versions or deliver different configurations.

This patch now reads the contents of `/etc/gentoo-release` to
distinguish between the two. We already do this is in a similar way with
the `/etc/redhat-release` to distinguish between CentOS, Scientific
Linux, RHEL, etc.

Because Gentoo und Funtoo are still similar and Funtoo claims to be
compatible with Gentoo upstream changes it may still be convenient to
have a fact to catch both operatingsystems. This can now be done with
the osfamily fact.